move context generator to better location
