+++ ohmg/frontend/views.py
@@ -9,7 +9,7 @@ from django.shortcuts import render, get_object_or_404
 from django.views import View
 from django.views.decorators.clickjacking import xframe_options_sameorigin
 
-from ohmg.core.context_processors import generate_ohmg_context
+from ohmg.core.http import generate_ohmg_context
 from ohmg.core.models import Map
 from ohmg.core.api.schemas import (
     LayerSetSchema,
